// types.ts
export const UserRole = {
  USER: 'user',
  EDITOR: 'editor',
  ADMIN: 'admin',
} as const;

export type UserRole = typeof UserRole[keyof typeof UserRole];

export interface User {
  id: string;
  email: string;
  username: string;
  role: UserRole;
  bio?: string;
  profilePictureUrl?: string; // URL to profile picture
  isSubscribed?: boolean; // For monetization
  following: string[]; // Array of author IDs the user is following
  // Do not store password on client-side for real user objects
}
// ...
export interface Comment {
  id: string;
  blogPostId: string;
  userId: string;
  userName: string; // Denormalized
  userProfilePictureUrl?: string; // Denormalized
  content: string;
  createdAt: string; // ISO Date string
  parentId?: string | null; // For replies
  replies?: Comment[]; // Nested replies
  reported?: boolean;
}